* Fixed the special double-bang replacement string for handling sensor name and value translations from using a comma to a colon as it was conflicting with split'ing the various variable/value pairs in the 'alerts' table. * Silenced a bunch of logs now that testing is well under way and I know they're not a current issue. * Cleaned up a bunch of formatting. * Fixed multiple bugs related to reporting IPMI sensor change alerts, as well as finished adding alerts to new and updated sensor values. More logic is needed for better messaging of non-temperature related changes, though. * Fixed a bug where celsius sensor values were being converted to fahrenheit, and the C/F suffix was inverted. * Changed all of the SQL schemas so that numeric data types (int, bigint and double precision) are now 'numerica', *except* for 'X_id' columns, so that floats with trailing 0s (ie: 1.250) don't lose the trailing zero when read back. This solves a problem where sensor values that had not changed between runs were being updated needlessly. * Fixed some SQL queries in scan-ipmitool to restrict queries/updates to the sensor's host. * Fixed a bug that was preventing entries to be added to the 'temperature' table. * Added a 'defined' check to some variables read in from the IPMI sensor data so that variables that didn't exist don't cause needless warnings to be printed to STDOUT. * Added the machine name to AN::Tools::Alert::check_alert_sent() method calls in scan-ipmitool to differentiate alerts from one node's sensors from the other node's sensors. * Added a buffer/hyeristics to temperature sensor thresholds to avoid repeated alerts when a sensor is floating on a threshold. Also made this user-definable. * Added timing logging to the actual IPMI calls. * Added 'percentage of free space'-based disk allocation to Striker's server provisioning. So now a user can provision a new server to use X% of the remaining free space in a pool, rather than always assigning a static value or using the 'all' keyword.

+++ ScanCore/ScanCore.sql
@@ -36,7 +36,7 @@ ALTER TABLE hosts OWNER TO #!variable!user!#;
 
 CREATE TABLE history.hosts (
 	history_id		bigserial,
-	host_id			bigint,
+	host_id			numeric,
 	host_name		text,
 	host_type		text,
 	host_emergency_stop	boolean				not null,
@@ -150,8 +150,8 @@ CREATE TABLE power (
 	power_record_locator	text,								-- Optional string used by the agent to identify the UPS
 	power_ups_fqdn		text				not null,			-- This is the full domain name of the UPS. This is used by ScanCore to determine which UPSes are powering a given node so this MUST match the host names used in the node's /etc/hosts file.
 	power_on_battery	boolean				not null,			-- TRUE == use "time_remaining" to determine if graceful power off is needed. FALSE == power loss NOT imminent, do not power off node. 
-	power_seconds_left	bigint,								-- Should always be set, but not required *EXCEPT* when 'power_on_battery' is TRUE.
-	power_charge_percentage	double precision,						-- Percentage charge in the UPS. Used to determine when the dashboard should boot the node after AC restore
+	power_seconds_left	numeric,								-- Should always be set, but not required *EXCEPT* when 'power_on_battery' is TRUE.
+	power_charge_percentage	numeric,						-- Percentage charge in the UPS. Used to determine when the dashboard should boot the node after AC restore
 	modified_date		timestamp with time zone	not null,
 	
 	FOREIGN KEY(power_host_id) REFERENCES hosts(host_id)
@@ -166,8 +166,8 @@ CREATE TABLE history.power (
 	power_record_locator	text,
 	power_ups_fqdn		text,
 	power_on_battery	boolean,
-	power_seconds_left	bigint,
-	power_charge_percentage	double precision,
+	power_seconds_left	numeric,
+	power_charge_percentage	numeric,
 	modified_date		timestamp with time zone	not null
 );
 ALTER TABLE history.power OWNER TO #!variable!user!#;
@@ -215,7 +215,7 @@ CREATE TABLE temperature (
 	temperature_host_id	bigint				not null,			-- The name of the node or dashboard that this temperature came from.
 	temperature_agent_name	text				not null,
 	temperature_sensor_name	text				not null,
-	temperature_celsius	double precision		not null,
+	temperature_celsius	numeric		not null,
 	temperature_state	text				not null,			-- warning, critical
 	temperature_is		text				not null,			-- high or low
 	modified_date		timestamp with time zone	not null,
@@ -230,7 +230,7 @@ CREATE TABLE history.temperature (
 	temperature_host_id	bigint,
 	temperature_agent_name	text,
 	temperature_sensor_name	text,
-	temperature_celsius	double precision,
+	temperature_celsius	numeric,
 	temperature_state	text,
 	temperature_is		text,
 	modified_date		timestamp with time zone	not null
@@ -338,7 +338,7 @@ CREATE TABLE ram_used (
 	ram_used_id		bigserial,
 	ram_used_host_id	bigint				not null,
 	ram_used_by		text				not null,			-- Either 'ScanCore' or the scan agent name
-	ram_used_bytes		bigint				not null,
+	ram_used_bytes		numeric				not null,
 	modified_date		timestamp with time zone	not null,
 	
 	FOREIGN KEY(ram_used_host_id) REFERENCES hosts(host_id)
@@ -350,7 +350,7 @@ CREATE TABLE history.ram_used (
 	ram_used_id		bigint,
 	ram_used_host_id	bigint,
 	ram_used_by		text				not null,			-- Either 'ScanCore' or the scan agent name
-	ram_used_bytes		bigint				not null,
+	ram_used_bytes		numeric				not null,
 	modified_date		timestamp with time zone	not null
 );
 ALTER TABLE history.ram_used OWNER TO #!variable!user!#;
